Fixes error where editing of large areas failed
An errors when exhausting the max length of URL parameters cause editing of
large areas to fail – for no appearent reason. This was corrected by changing
all geometry-carrying GET requests to POST requests by removing their relience
on JSONP.

This requires the page to be allowed by remote origin policy or else this method
will fail (since we can not use JSONP).